Allow ip from specified network to be used
It's not unusual to find a container connected to more than one network.
This change allows us to specify the network to be used via a container
label, "com.dnsdock.network".

A couple things to note:

No, there is no equivalent environment variable that can be set.
(Setting meta-configuration via the container _environment_ just seems
wrong, frankly.)

We handle the label usage in a different place and manner than the
others, as, oddly, it seems like the least invasive way to do it.
(Right now, this is a plus.)  The "override" function only has access to
the labels themselves, and approaches things from a "use these literals
instead" perspective -- neither of which we want here.