Use compose labels if present
dnsdock has a way of determining what a containers fqdn should be
involving the actual container name, image name, etc.  However, if the
compose labels are present, we can certainly use them to generate a
sensible hostname without needing to mess with things that shouldn't be
configuration, like image names.  (Not to be opinionated or anything.
:))